# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AIADR API
  version: 0.1.0.dev0
  description: Contract for the local AI-Assisted Data Review application.
servers:
  - url: /api/v1
...
paths:
...
  /instruction-sets:
    get:
      tags: [instruction-sets]
      operationId: listInstructionSets
      responses:
        "200":
          description: Valid portable instruction sets available to new sessions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstructionSetsResponse"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [instruction-sets]
      operationId: createInstructionSet
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InstructionSetCreateRequest" }
      responses:
        "201":
          description: ID of the created instruction set.
          content:
            application/json:
              schema: { type: string }
        default:
          $ref: "#/components/responses/Error"
  /instruction-sets/{instruction_set_id}:
    parameters:
      - name: instruction_set_id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [instruction-sets]
      operationId: getInstructionSet
      responses:
        "200":
          description: Editable instruction-set definition.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InstructionSetDefinition" }
        default:
          $ref: "#/components/responses/Error"
    put:
      tags: [instruction-sets]
      operationId: updateInstructionSet
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InstructionSetUpdateRequest" }
      responses:
        "204":
          description: Instruction set replaced.
        default:
          $ref: "#/components/responses/Error"
    delete:
      tags: [instruction-sets]
      operationId: deleteInstructionSet
      responses:
        "204":
          description: Instruction set removed from the catalogue.
        default:
          $ref: "#/components/responses/Error"
  /instruction-sets/{instruction_set_id}/active:
    parameters:
      - name: instruction_set_id
        in: path
        required: true
        schema: { type: string }
    put:
      tags: [instruction-sets]
      operationId: activateInstructionSet
      responses:
        "204":
          description: Instruction set selected for future analysis.
        default:
          $ref: "#/components/responses/Error"
...
  /sessions/{session_id}/model-log:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [sessions]
      operationId: getModelLogs
      description: >-
        Return model-call summaries. In development mode, debug fields may
        additionally contain sensitive request, response, and provider diagnostic
        text.
      responses:
        "200":
          description: Model-call summaries and optional development diagnostics.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelLogsResponse"
        default:
          $ref: "#/components/responses/Error"
  /sessions/{session_id}/analysis:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    post:
      tags: [sessions]
      operationId: requestAnalysis
      responses:
        "202":
          description: Analysis request accepted.
        default:
          $ref: "#/components/responses/Error"
  /sessions/{session_id}/analysis/cancel:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    post:
      tags: [sessions]
      operationId: cancelAnalysis
      responses:
        "202":
          description: Analysis cancellation requested.
        "204":
          description: No active analysis.
        default:
          $ref: "#/components/responses/Error"
  /sessions/{session_id}/findings:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    post:
      tags: [sessions]
      operationId: createManualFinding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualFindingCreateRequest"
      responses:
        "201":
          description: Generated layer ID for the created manual finding.
          content:
            application/json:
              schema:
                type: string
        default:
          $ref: "#/components/responses/Error"
  /sessions/{session_id}/findings/{finding_id}:
    parameters:
      - $ref: "#/components/parameters/SessionId"
      - $ref: "#/components/parameters/FindingId"
    patch:
      tags: [sessions]
      operationId: updateFinding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FindingUpdateRequest"
      responses:
        "204":
          description: Finding updated.
        default:
          $ref: "#/components/responses/Error"
  /sessions/{session_id}/classification-options:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [sessions]
      operationId: getClassificationOptions
      responses:
        "200":
          description: Entity types available from the session instruction set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassificationOptionsResponse"
        default:
          $ref: "#/components/responses/Error"
  /sessions/{session_id}/layers:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [sessions]
      operationId: getLayers
      responses:
        "200":
          description: Hydrated review layers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LayersResponse"
        default:
          $ref: "#/components/responses/Error"
  /sessions/{session_id}/layers/{layer_id}:
    parameters:
      - $ref: "#/components/parameters/SessionId"
      - $ref: "#/components/parameters/LayerId"
    put:
      tags: [sessions]
      operationId: updateLayer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LayerUpdateRequest"
      responses:
        "204":
          description: Layer updated.
        default:
          $ref: "#/components/responses/Error"
  /sessions/{session_id}/layers/{layer_id}/effect-override:
    parameters:
      - $ref: "#/components/parameters/SessionId"
      - $ref: "#/components/parameters/LayerId"
    delete:
      tags: [sessions]
      operationId: resetEffectOverride
      responses:
        "204":
          description: Layer restored to its instruction-set policy effect.
        default:
          $ref: "#/components/responses/Error"
...
  /settings/models:
    get:
      tags: [settings]
      operationId: listModels
      responses:
        "200":
          description: Saved models and active selection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelsResponse"
        default:
          $ref: "#/components/responses/Error"
  /settings/dependencies:
    get:
      tags: [settings]
      operationId: getDependencies
      responses:
        "200":
          description: Current availability of local media dependencies.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DependenciesResponse"
        default:
          $ref: "#/components/responses/Error"
  /settings/models/{model_id}:
    parameters:
      - $ref: "#/components/parameters/ModelId"
    put:
      tags: [settings]
      operationId: updateModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelWriteRequest"
      responses:
        "204":
          description: Model created or updated.
        default:
          $ref: "#/components/responses/Error"
    delete:
      tags: [settings]
      operationId: deleteModel
      responses:
        "204":
          description: Model deleted.
        default:
          $ref: "#/components/responses/Error"
  /settings/models/{model_id}/model-metadata:
    parameters:
      - $ref: "#/components/parameters/ModelId"
    get:
      tags: [settings]
      operationId: getModelModelMetadata
      responses:
        "200":
          description: Provider-reported metadata for the configured model.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelMetadata"
        default:
          $ref: "#/components/responses/Error"
  /settings/models/{model_id}/active:
    parameters:
      - $ref: "#/components/parameters/ModelId"
    put:
      tags: [settings]
      operationId: activateModel
      responses:
        "204":
          description: Model selected for future analysis.
        default:
          $ref: "#/components/responses/Error"
  /settings/review-options:
    get:
      tags: [settings]
      operationId: getReviewOptions
      responses:
        "200":
          description: Review decisions and supported redaction effects.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewOptions"
        default:
          $ref: "#/components/responses/Error"
components:
  parameters:
    SessionId:
      name: session_id
      in: path
      required: true
      schema: { type: string }
    FindingId:
      name: finding_id
      in: path
      required: true
      schema: { type: string }
    LayerId:
      name: layer_id
      in: path
      required: true
      schema: { type: string }
    ModelId:
      name: model_id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Error:
      description: Structured application error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorPayload:
      type: object
      additionalProperties: false
      required: [code, message, category, retryable, details, correlation_id]
      properties:
        code: { $ref: "#/components/schemas/ErrorCode" }
        message: { type: string }
        category: { $ref: "#/components/schemas/ErrorCategory" }
        retryable: { type: boolean }
        details:
          type: [object, "null"]
          additionalProperties: true
        correlation_id: { type: string }
    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorPayload"
    SourceKind:
      type: string
      enum: [document, image, audio]
...
    InstructionSetSummary:
      type: object
      additionalProperties: false
      required: [id, name, content_hash]
      properties:
        id: { type: string }
        name: { type: string }
        content_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
    InstructionSetsResponse:
      type: object
      additionalProperties: false
      required: [items, active_instruction_set_id]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/InstructionSetSummary" }
        active_instruction_set_id: { type: [string, "null"] }
    InstructionSetPrompts:
      type: object
      additionalProperties: false
      required: [text, document, image, audio]
      properties:
        text: { type: string, minLength: 1 }
        document: { type: string, minLength: 1 }
        image: { type: string, minLength: 1 }
        audio: { type: string, minLength: 1 }
    InstructionSetDefinition:
      type: object
      additionalProperties: false
      required: [id, name, content_hash, policy, prompts]
      properties:
        id: { type: string }
        name: { type: string }
        content_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        policy:
          type: object
          additionalProperties: true
        prompts: { $ref: "#/components/schemas/InstructionSetPrompts" }
    InstructionSetCreateRequest:
      type: object
      additionalProperties: false
      required: [id, name, policy, prompts]
      properties:
        id: { type: string }
        name: { type: string }
        policy:
          type: object
          additionalProperties: true
        prompts: { $ref: "#/components/schemas/InstructionSetPrompts" }
    InstructionSetUpdateRequest:
      type: object
      additionalProperties: false
      required: [name, expected_content_hash, policy, prompts]
      properties:
        name: { type: string }
        expected_content_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        policy:
          type: object
          additionalProperties: true
        prompts: { $ref: "#/components/schemas/InstructionSetPrompts" }
...
    ReviewDecision:
      type: string
      enum: [needs_review, confirmed, preserved]
...
    FindingTarget:
      oneOf:
        - $ref: "#/components/schemas/ImageTarget"
        - $ref: "#/components/schemas/DocumentTarget"
        - $ref: "#/components/schemas/AudioTarget"
      discriminator:
        propertyName: kind
        mapping:
          image: "#/components/schemas/ImageTarget"
          document: "#/components/schemas/DocumentTarget"
          audio: "#/components/schemas/AudioTarget"
...
    LayerAction:
      type: string
      enum: [redact, mask, pseudonymize, preserve]
    LayerEffect:
      type: string
      enum: [box, blur, pixelate, partial_mask, token_replace, bleep, mute, none]
    EffectSource:
      type: string
      enum: [policy, reviewer]
    Layer:
      type: object
      additionalProperties: false
      required: [id, finding, action, effect, effect_source, enabled, fill_color, custom_text, note]
      properties:
        id: { type: string }
        finding: { $ref: "#/components/schemas/Finding" }
        action: { $ref: "#/components/schemas/LayerAction" }
        effect: { $ref: "#/components/schemas/LayerEffect" }
        effect_source: { $ref: "#/components/schemas/EffectSource" }
        enabled: { type: boolean }
        fill_color: { type: string }
        custom_text: { type: [string, "null"] }
        note: { type: [string, "null"] }
    LayersResponse:
      type: object
      additionalProperties: false
      required: [session_id, layers]
      properties:
        session_id: { type: string }
        layers:
          type: array
          items: { $ref: "#/components/schemas/Layer" }
    ManualFindingCreateRequest:
      type: object
      additionalProperties: false
      required: [label, reviewed_entity_type, description, reason, target]
      properties:
        label: { type: string }
        reviewed_entity_type: { type: string }
        description: { type: string }
        reason: { type: string }
        target: { $ref: "#/components/schemas/FindingTarget" }
    FindingUpdateRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        label: { type: string }
        reviewed_entity_type: { type: [string, "null"] }
        description: { type: [string, "null"] }
        reason: { type: [string, "null"] }
        review_decision: { $ref: "#/components/schemas/ReviewDecision" }
        reviewer_note: { type: [string, "null"] }
        target: { $ref: "#/components/schemas/FindingTarget" }
    LayerUpdateRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        action: { $ref: "#/components/schemas/LayerAction" }
        effect: { $ref: "#/components/schemas/LayerEffect" }
        enabled: { type: boolean }
        fill_color: { type: string }
        custom_text: { type: [string, "null"] }
        note: { type: [string, "null"] }
...
    DependencyStatus:
      type: object
      additionalProperties: false
      required: [available, path, version, error]
      properties:
        available: { type: boolean }
        path: { type: [string, "null"] }
        version: { type: [string, "null"] }
        error: { type: [string, "null"] }
    DependenciesResponse:
      type: object
      additionalProperties: false
      required: [ffmpeg, libreoffice]
      properties:
        ffmpeg: { $ref: "#/components/schemas/DependencyStatus" }
        libreoffice: { $ref: "#/components/schemas/DependencyStatus" }
    ApiFormat:
      type: string
      enum: [openai_compatible, anthropic_messages, google_genai]
...
    ModelCapabilities:
      type: object
      additionalProperties: false
      required: [supports_vision, supports_audio, supports_json_mode]
      properties:
        supports_vision: { type: boolean }
        supports_audio: { type: boolean }
        supports_json_mode: { type: boolean }
...
    ProviderWriteSettings:
      oneOf:
        - $ref: "#/components/schemas/OpenAICompatibleWriteSettings"
        - $ref: "#/components/schemas/AnthropicMessagesWriteSettings"
        - $ref: "#/components/schemas/GoogleGenAIWriteSettings"
      discriminator:
        propertyName: api_format
        mapping:
          openai_compatible: "#/components/schemas/OpenAICompatibleWriteSettings"
          anthropic_messages: "#/components/schemas/AnthropicMessagesWriteSettings"
          google_genai: "#/components/schemas/GoogleGenAIWriteSettings"
...
    ModelWriteRequest:
      type: object
      additionalProperties: false
      required: [label, provider, capabilities]
      properties:
        label: { type: string, minLength: 1, pattern: ".*\\S.*" }
        provider: { $ref: "#/components/schemas/ProviderWriteSettings" }
        capabilities: { $ref: "#/components/schemas/ModelCapabilities" }
    Model:
      type: object
      additionalProperties: false
      required: [model_id, label, provider, capabilities, configured]
      properties:
        model_id: { type: string }
        label: { type: string, minLength: 1, pattern: ".*\\S.*" }
        provider: { $ref: "#/components/schemas/ProviderModelSettings" }
        capabilities: { $ref: "#/components/schemas/ModelCapabilities" }
        configured: { type: boolean }
    ModelMetadata:
      type: object
      additionalProperties: true
      description: Provider-native model metadata represented as JSON.
    ModelsResponse:
      type: object
      additionalProperties: false
      required: [models, active_model_id]
      properties:
        models:
          type: array
          items: { $ref: "#/components/schemas/Model" }
        active_model_id: { type: [string, "null"] }
    ReviewOptions:
      type: object
      additionalProperties: false
      required: [review_decisions, supported_effects_by_kind_and_action]
      properties:
        review_decisions:
          type: array
          items: { $ref: "#/components/schemas/ReviewDecision" }
        supported_effects_by_kind_and_action:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: array
              items: { $ref: "#/components/schemas/LayerEffect" }
    ClassificationOption:
      type: object
      additionalProperties: false
      required: [entity_type, display_name, description]
      properties:
        entity_type: { type: string }
        display_name: { type: string }
        description: { type: [string, "null"] }
    ClassificationOptionsResponse:
      type: object
      additionalProperties: false
      required: [options]
      properties:
        options:
          type: array
          items: { $ref: "#/components/schemas/ClassificationOption" }
    ModelInteractionStatus:
      type: string
      enum: [started, succeeded, failed, cancelled]
    TokenCountMethod:
      type: string
      enum: [provider, estimated]
...
    ModelRequestSummary:
      type: object
      additionalProperties: false
      required: [source_payload_bytes, line_count, character_count, attachment_kind, attachment_mime_type, attachment_size_bytes, attachment_sha256, system_prompt_sha256]
      properties:
        source_payload_bytes: { type: integer, minimum: 0 }
        line_count: { type: [integer, "null"], minimum: 0 }
        character_count: { type: [integer, "null"], minimum: 0 }
        attachment_kind:
          oneOf:
            - $ref: "#/components/schemas/SourceKind"
            - type: "null"
        attachment_mime_type: { type: [string, "null"] }
        attachment_size_bytes: { type: [integer, "null"], minimum: 0 }
        attachment_sha256:
          type: [string, "null"]
          pattern: "^[0-9a-f]{64}$"
        system_prompt_sha256:
          type: string
          pattern: "^[0-9a-f]{64}$"
...
    DebugModelIO:
      type: object
      description: Development-only textual provider diagnostics. Binary attachments are omitted.
      additionalProperties: false
      required: [request_payload, response_content, error_message]
      properties:
        request_payload:
          type: object
          additionalProperties: true
        response_content: { type: string }
        error_message: { type: [string, "null"] }
    ModelInteractionLog:
      type: object
      description: Content-minimized model-call metadata with optional development diagnostics.
      additionalProperties: false
      required: [log_id, session_id, created_at, status, completed_at, duration_ms, kind, page, api_format, model, request_summary, result_summary, debug, finish_reason, input_count_method, estimated_input_tokens, provider_counted_input_tokens, requested_output_tokens, max_input_tokens, actual_input_tokens, actual_output_tokens, total_tokens, reasoning_tokens]
      properties:
        log_id: { type: string }
        session_id: { type: string }
        created_at: { type: string, format: date-time }
        status: { $ref: "#/components/schemas/ModelInteractionStatus" }
        completed_at: { type: [string, "null"], format: date-time }
        duration_ms: { type: [integer, "null"], minimum: 0 }
        kind: { $ref: "#/components/schemas/SourceKind" }
        page: { type: [integer, "null"], minimum: 1 }
        api_format: { $ref: "#/components/schemas/ApiFormat" }
        model: { type: string }
        request_summary: { $ref: "#/components/schemas/ModelRequestSummary" }
        result_summary: { $ref: "#/components/schemas/ModelResultSummary" }
        debug:
          oneOf:
            - $ref: "#/components/schemas/DebugModelIO"
            - type: "null"
        finish_reason: { type: [string, "null"] }
        input_count_method:
          oneOf:
            - $ref: "#/components/schemas/TokenCountMethod"
            - type: "null"
        estimated_input_tokens: { type: [integer, "null"], minimum: 0 }
        provider_counted_input_tokens: { type: [integer, "null"], minimum: 0 }
        requested_output_tokens: { type: [integer, "null"], minimum: 0 }
        max_input_tokens: { type: [integer, "null"], minimum: 0 }
        actual_input_tokens: { type: [integer, "null"], minimum: 0 }
        actual_output_tokens: { type: [integer, "null"], minimum: 0 }
        total_tokens: { type: [integer, "null"], minimum: 0 }
        reasoning_tokens: { type: [integer, "null"], minimum: 0 }
    ModelLogsResponse:
      type: object
      additionalProperties: false
      required: [session_id, logs]
      properties:
        session_id: { type: string }
        logs:
          type: array
          items: { $ref: "#/components/schemas/ModelInteractionLog" }
...